example/requirements.txt
# Your app requirements.
-r ../requirements_test.txt
# Your app in editable mode.
-e ../

django-autofixture==0.12.1
